Final version of the notebook
The objective of the project is to analyse the "Speaker Accent Recognition Dataset" from the UCI Machine Learning Repository. The task is to determine if the dataset is linearly separable in the feature space defined by the linear Principal Component Analysis (PCA) or in the feature space defined by the non-linear PCA with the use of the Radial Basis Function Kernel.